% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_drug.R
\name{clean_drug}
\alias{clean_drug}
\title{clean up date and type of treatment for one type of drug}
\usage{
clean_drug(trs, trs_date, drugname)
}
\arguments{
\item{trs}{A dataframe that only includes the treatment medication}

\item{trs_date}{A dataframe that only includes the treatment dates.}

\item{durgname}{durgnames(can be multiple)}
}
\value{
A list includes three elements: whether the drug is included, that start date and end date of the treatment
}
\description{
This function take multiple lines of treatment, identify which patient has a certain drug, output the drug name. start date and end date of this line of treatment
}
\examples{
drugname1 = c("erlotinib","gefitinib")
trlinesd   = c("tr2","tr3","tr4","tr5")
trlines_datesd   = c("tr2date","tr3date","tr4date","tr5date")
trs <- Dat[trlines]
trs_date <- Dat[trlines_dates]
cdrug1 <- clean_drug(trs,trs_date,drugname1)
}
