% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_drugs.R
\name{clean_drugs}
\alias{clean_drugs}
\title{clean up date and type of treatment}
\usage{
clean_drugs(
  Dat,
  drugname1,
  drugname2,
  trlines,
  trlines_dates,
  drugname3 = NULL,
  drugDup = FALSE
)
}
\arguments{
\item{Dat}{A data.frame includes all the data}

\item{trlines}{A vector that includes all treatment variable names}

\item{trlines_dates}{A vector that includes all treatment date varaible names}

\item{drugDup}{a boolean indicates if duplicate treatment for a patients is allowed}

\item{durgname1}{durgnames(can be multiple)}

\item{durgname2}{durgnames(can be multiple)}
}
\value{
A list includes three elements: drug group, that start date and end date of the treatment
}
\description{
This function take multiple lines of treatment, identify which patient has wanted drugs, output the drug name. start date and end date of this line of treatment
}
\examples{
drugname1 = c("erlotinib","gefitinib")
drugname2 = c("osimertinib")
trlinesd   = c("tr2","tr3","tr4","tr5")
trlines_datesd   = c("tr2date","tr3date","tr4date","tr5date")
tr_info <- clean_drugs(D, drugname1, drugname2,trlines = trlinesd, trlines_dates = trlines_datesd)
}
