% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crisk_cat.R
\name{crisk_cat}
\alias{crisk_cat}
\title{A output table for univariable risk analysis (catigorical only)}
\usage{
crisk_cat(csurv, cevent, cvars, gnames, month = 0, y1 = T, y2 = T, y5 = T)
}
\arguments{
\item{csurv}{Follow up time}

\item{cevent}{The status indicator, normally 0 = alive, 1 = event, 2 = other event}

\item{cvars}{one Catagorical Variable}

\item{gnames}{A text vector of the the group name for output}
}
\value{
Output: a well-formated table , with N , N of event. , N of competing event, 1-year, 2-year, 5-year, n- year incidence rate, Hazard ratio , P, c-index
}
\description{
# Output: a well-formated table , with N , N of event. , N of competing event,
}
\examples{
output <- crisk_cat(NewDat$pfs, NewDat$dcensor, NewDat$Histologic.type, "gnames")

}
