% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crisk_con.R
\name{crisk_con}
\alias{crisk_con}
\title{output for univariable and multivariable competing risk analysis (continuous and ordinal only)}
\usage{
crisk_con(csurv, cevent, cvars, gnames)
}
\arguments{
\item{csurv}{Follow up time}

\item{cevent}{The status indicator, normally 0 = alive, 1 = event, 2 = other event}

\item{cvars}{A Matrix of groups variables (continuous and ordinal only)}

\item{gnames}{A text vector of the the group name for output}
}
\value{
A dataframed output including HRs (95\% Confidence Intervals), P values.
}
\description{
output the table with general competing risk analysis result with HR (95\% Confidence Interval),P value. This function only change the format of the output.
}
\examples{
X      <- cbind(D$v1, D$v2 , D$age_ge60)
Gnames <- c('SRSF2/U2FA1(WT vs Mutation)', 'Disease(AML vs MPN)', 'Age(< 60 vs > 60)')
output <- crisk_con(D$surv, D$censor_rm, X, Gnames)

}
