% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor2ind.R
\name{factor2ind}
\alias{factor2ind}
\title{Make An Integer Matrix Out of A Factor Variable.}
\usage{
factor2ind(x, baseline)
}
\arguments{
\item{x}{a variable.}

\item{baseline}{a string indicating the reference level.}
}
\value{
a matrix
}
\description{
Create an indicator matrix of dimension length(x) x (nlevels(x)-1) with the column corresponding
to the baseline level removed (by default the first level is used as baseline).
}
\examples{
x = gl(4, 2, labels = c( "A", "B", "C", "D"))
factor2ind(x)
factor2ind(x, "C")
}
