% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findjava.R
\name{findjava}
\alias{findjava}
\title{find java path , solve the rjava problem}
\usage{
findjava()
}
\value{
all of your java paths
}
\description{
find java path , solve the rjava problem
}
\examples{
findjava
Sys.setenv(JAVA_HOME='C:\\\\Your\\\\Java\\\\Directory')
library(rJava)
}
