% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixchar.R
\name{fixchar}
\alias{fixchar}
\title{clearn up the all the char variable in the dataset , return a dataset}
\usage{
fixchar(Data)
}
\arguments{
\item{Data}{Dataset}
}
\value{
Output: a well-formated table , with N , N of event. , N of competing event, 1-year, 2-year, 5-year, n- year incidence rate, Hazard ratio , P, c-index
}
\description{
clearn up the all the char variable in the dataset , return a dataset
}
\examples{
D <- fixchar(D)

}
