% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tr_pfs.R
\name{get_tr_pfs}
\alias{get_tr_pfs}
\title{get pfs variable when multiple end dates excists}
\usage{
get_tr_pfs(start_date, end_date, other_dates, date_death, date_fup)
}
\arguments{
\item{start_date}{vector, start date for a treatment}

\item{end_date}{vector, end date for the treatment}

\item{other_dates}{data frame. other protencial progression dates, eg srs}

\item{date_death}{vector}

\item{date_fup}{vector}
}
\value{
A list includes three elements: progression dates, progression sensor variable , pfs
}
\description{
Take the last date for treatment, all other dates, date of death and last follow up and return pfs ttf variable
}
\examples{
cod <-  c("date.srs1","date.srs2","date.srs3","date.srs4","date.srs5","date.srs6")
DT_other_datas <- D[cod]
EGFR25_pfs <- get_tr_pfs(D$date_25start, EGFR25_info[[3]],DT_other_datas,D$date_death, D$date_lastfollowup)
}
