% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingorfactor_c.R
\name{ingorfactor_c}
\alias{ingorfactor_c}
\title{A helper fucntion for competing risk functions}
\usage{
ingorfactor_c(x, event)
}
\arguments{
\item{x}{Variable need to be cleaned}

\item{Event}{The status indicator, normally 0=censored, 1=event, 2 = death}
}
\value{
x variable with some NA Values
}
\description{
for categorical data, will condiser groups < 10 without event as NA, Since the crr function will fail to run in this case
}
\examples{
ingorfactor_c(NewDat$Race, NewDat$dcensor)
}
