% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jdigit.R
\name{J.digit}
\alias{J.digit}
\title{Formatting Decimal places}
\usage{
J.digit(x, k)
}
\arguments{
\item{x}{Number needed to be formated}

\item{k}{the number of decimals to show.}
}
\value{
rounded number with k decimals
}
\description{
J.digit format the numbers
}
\examples{
specify_decimal(1234, 5)
"1234.00000"
specify_decimal(0.1234, 5)
"0.12340"
}
