% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JB.logreg.R
\name{JB.logreg}
\alias{JB.logreg}
\title{A complete table for univariable survival analysis}
\usage{
JB.logreg(Dat, name, yname)
}
\arguments{
\item{Dat}{A data.frame that includes two column x and y}

\item{name}{X variable output name}

\item{yname}{Y variable noutput name}
}
\description{
JB.logreg return results for logistic regression: Varname, levels, outcome(yes),outcome(no), OR , P
}
\examples{
colnames(Dat) <- c("x","y")
rs <- JB.logreg(Dat, name[i],yname)
}
