% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JB.logreg_allatOnce.R
\name{JB.logreg_allatOnce}
\alias{JB.logreg_allatOnce}
\title{A output table for univariable and multivariable logistic regression}
\usage{
JB.logreg_allatOnce(D, xvar, yvar, xname, yname)
}
\arguments{
\item{D}{Dataset}

\item{xvar}{a vector of X variable colnames}

\item{yvar}{Y variable colname}

\item{xname}{X output label}

\item{yname}{Y output label}
}
\description{
JB.logreg_allatOnce  return results for UMA and  MVA logistic regression in a list
}
\examples{
f1 = glm(as.matrix(y~.,family = binomial, data = tmp)
rs <- JB.logreg_multi(f1)
}
