% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JB.logreg_m.R
\name{JB.logreg_m}
\alias{JB.logreg_m}
\title{A complete table for univariable survival analysis}
\usage{
JB.logreg_m(xvar, yvar, name, yname)
}
\arguments{
\item{xvar}{a dataframe has all the X variables}

\item{yvar}{a vector of outcome variable}

\item{name}{X variable output name}

\item{yname}{Y variable noutput name}
}
\description{
JB.logreg_m return logistic regression  results for for each of the input variables : Varname, evels, outcome(yes),outcome(no),  OR , P
}
\examples{
rs <- JB.logreg_m(Date, D$Hypertension10, cats[1:16],"Hypertension")
}
