% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JB.logreg_multi.R
\name{JB.logreg_multi}
\alias{JB.logreg_multi}
\title{A output table for multivariable logistic regression}
\usage{
JB.logreg_multi(outx)
}
\arguments{
\item{outx}{glm returning result}
}
\description{
JB.logreg_multi return results for MVA logistic regression: Varname,OR , P
}
\examples{
f1 = glm(as.matrix(y~.,family = binomial, data = tmp)
rs <- JB.logreg_multi(f1)
}
