% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JSinsert.R
\name{JS.insert}
\alias{JS.insert}
\title{Insert rows into dataset}
\usage{
JS.insert(locs, inrows, df, rep = T)
}
\arguments{
\item{locs}{Row locations to insert the row}

\item{inrows}{Rows that need to be inserted}

\item{df}{Dataset}
}
\value{
A dataset with inserted rows
}
\description{
JS.insert inserts rows into dataset
}
\examples{
aa <- data.frame(1:6, col = 2)
a <- c(7,7)
JS.insert (c(1,2),c(7,7),aa) 


}
