% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JS.multi.R
\name{JS.multi}
\alias{JS.multi}
\title{A output table for multivariable survival analysis}
\usage{
JS.multi(...)
}
\arguments{
\item{...}{arguments will be passed to coxph}
}
\value{
A dataframe of coxph output including Variable names,  HRs (95\% Confidence Intervals), P values
}
\description{
JS.multi output the table with general multivariable survival analysis result with  HR (95\% Confidence Interval),P value
}
\examples{
JS.multi (Surv(as.numeric(pd_surv), pd_censor) ~ as.factor(tr_group) + as.factor(BMI_c) + BMI, data = D)


}
