% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JSp.R
\name{JS.p}
\alias{JS.p}
\title{Format P value}
\usage{
JS.p(value)
}
\arguments{
\item{value}{numeric vector of p-value (possibly with NAs )}
}
\value{
A numeric vector of corrected p-value
}
\description{
JS.p is intended for formatting p-value
}
\examples{
JS.p(0.060)
JS.p(0.0050)
JS.p(0.00001)
}
