% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p2excel.R
\name{p2excel}
\alias{p2excel}
\title{Export the A Single Dataframe to An Excel Sheet}
\usage{
p2excel(
  tabname = "Default",
  datastable,
  tablename = "Default",
  filename = "Default.xlsx"
)
}
\arguments{
\item{tabname}{a string with the tab's name.}

\item{datastable}{the dataframe that will be exported to Excel.}

\item{tablename}{a string containing the table label and title, which will appear as the first row}

\item{filename}{the name of the spreadsheet}
}
\value{
a spreadsheet containing an exported tables
}
\description{
The function saves a dataframe into an excel sheet with a predetermined format.
}
