% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_select_model.R
\name{random_select_model}
\alias{random_select_model}
\alias{JS.multi_modify}
\title{A output table for multivariable survival analysis}
\usage{
JS.multi_modify(data, multi_result)
}
\arguments{
\item{Data}{A data.frame in which to interpret the variables}

\item{Event}{The status indicator, normally 0=alive, 1=dead}

\item{Stime}{This is the follow up time}

\item{Svar}{A vector of group}
}
\value{
A Matrix of JS.multi
}
\description{
random_select_model use random forest for variable selection, then push top variables into JS.multi
}
\examples{
Svar <- c(cat,"first.line.systemic.therapies","age","SRS_total_number")
ttt <- random_select_model(D2,"os_month_srs1", "scensor",Svar,"first.line.systemic.therapies")

}
