% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_multi.R
\name{surv_multi}
\alias{surv_multi}
\title{Modify the Output for a Multi-variable Survival Analysis.}
\usage{
surv_multi(...)
}
\arguments{
\item{...}{arguments will be passed to coxph}
}
\value{
a dataframe containing coxph output that includes variable names, HRs (95% CIs), and P values.
}
\description{
Create a table with the general multi-variable survival analysis results, including the HR (95 percent CI), P value.
}
\examples{
Dat <- survival::lung
surv_multi(survival::Surv(time, status) ~ as.factor(sex) + age + meal.cal, data = Dat)

}
