% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crisk_multi.R
\name{crisk_multi}
\alias{crisk_multi}
\title{A output table for multivariable competing risk model}
\usage{
crisk_multi(data, csurv, cevent, mvars, gnames, con_var = NULL, cat_var = NULL)
}
\arguments{
\item{data}{Dataset}

\item{csurv}{Follow up time}

\item{cevent}{The status indicator, normally 0 = alive, 1 = event, 2 = other event}

\item{mvars}{one Catagorical Variable}

\item{gnames}{A text vector of the the group name for output}
}
\value{
Output: a well-formated table ,  Hazard ratio , P
}
\description{
# Output:  Hazard ratio , P, c-index
}
\examples{
mvars <- c("age50", "Race", "Sentinel.lymph.node.positivity", "Surgical.margin.positivity", "Vascular.invasion", "Surgery.type", "Radiation.therapy.received", "Neoadjuvant.Chemotherapy" , "Adjuvant.chemotherapy","Overall.clincal.staging..TNM." )
multi_out_d<- crisk_multi(NewDat2, "pfs","dcensor",mvars)

}
