% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crisk_multiuni.R
\name{crisk_multiuni}
\alias{crisk_multiuni}
\title{A complete table for univariable compeiting risk analysis}
\usage{
crisk_multiuni(
  Data,
  Stime,
  Event,
  catSvars = NULL,
  conSvars = NULL,
  ordSvars = NULL,
  y1 = T,
  y2 = T,
  y5 = T,
  month = 0
)
}
\arguments{
\item{Data}{A data.frame in which to interpret the variables}

\item{Stime}{This is the follow up time}

\item{Event}{The status indicator, normally 0=alive, 1=dead}

\item{catSvars}{A vector of cat variables}

\item{conSvars}{A vector of con variables}

\item{ordSvars}{A vector of ordinal variables}

\item{y1}{1-year survival rate (yes/no)}

\item{y2}{2-year survival rate(yes/no)}

\item{y5}{5-year survival rate(yes/no)}
}
\value{
A tibble of survival output
}
\description{
crisk_multiuni output the table with general survival analysis result with Number of total patients,
Number of Events, Estimated Median, 1,2,5 year rate, HR (95\% Confidence Interval),P value This function only change the format of the output table.
}
\examples{
con_var <- colnames(NewDat)[sapply(NewDat, class) == "numeric"][c(-1, -(6:11))]
ord_var <- colnames(NewDat)[sapply(NewDat, class) == "factor"]
cat_var <- colnames(NewDat)[sapply(NewDat, class) == "character"][c(-2, -8,-23, -25, -26, -c(10:17))]
uni_out <- crisk_multiuni(NewDat, "pfs", "dcensor", catSvars = cat_var, ordSvars = ord_var, conSvars =con_var)
}
