% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crisk_ord.R
\name{crisk_ord}
\alias{crisk_ord}
\title{A output table for univariable and multivariable competing risk analysis (ordinal only)}
\usage{
crisk_ord(csurv, cevent, cvars, gnames, month = 0, y1 = T, y2 = T, y5 = T)
}
\arguments{
\item{csurv}{Follow up time}

\item{cevent}{The status indicator, normally 0 = alive, 1 = event, 2 = other event}

\item{cvars}{A Matrix of groups variables (continuous and ordinal only)}

\item{gnames}{A text vector of the the group name for output}
}
\value{
Output: a well-formated table , with N , N of event. , N of competing event, 1-year, 2-year, 5-year, n- year incidence rate, Hazard ratio , P, c-index
}
\description{
# Output: a well-formated table , with N , N of event. , N of competing event,
}
\examples{
output <- crisk_ord(NewDat$pfs, NewDat$dcensor, NewDat$ECOG.Score, "gnames")

}
