% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p2excel.R
\name{p2excel}
\alias{p2excel}
\title{The excel output of a dataset}
\usage{
p2excel(
  tabname = "Default",
  datastable,
  tablename = "Default",
  filename = "Default.xlsx"
)
}
\arguments{
\item{tabname}{The name of the tab,}

\item{datastable}{The output dataset}

\item{tablename}{The name of the table as the frist row}

\item{filename}{The name of the excel file}
}
\value{
An excel sheet that ouputs the table
}
\description{
The excel output of a dataset
}
\examples{
p2excel(tabname = "Uni PFS", datastable = uni_out, tablename = "Table 2, Univariable PFS", filename = "Table2.xlsx" )
p2excel(datastable = uni_out)
}
