% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p2excel_pre.R
\name{p2excel_pre}
\alias{p2excel_pre}
\title{The excel output of a dataset prepration}
\usage{
p2excel_pre(tabname = "Default", datastable, tablename = "Default",
  filesave)
}
\arguments{
\item{tabname}{The name of the tab,}

\item{datastable}{The output dataset}

\item{tablename}{The name of the table as the frist row}

\item{filename}{The name of the excel file}
}
\value{
An excel sheet that ouputs the table
}
\description{
The excel output of a dataset prepration
}
\examples{
wb <- createWorkbook()
wb <- p2excel_pre(tabname = "Uni PFS", datastable = uni_out, tablename = "Table 2, Univariable PFS", filesave = wb)
saveWorkbook(wb, file = "filename.xlsx", overwrite = TRUE)
}
