% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_multiuni.R
\name{surv_multiuni}
\alias{surv_multiuni}
\title{A complete table for univariable survival analysis}
\usage{
surv_multiuni(
  Data,
  Stime,
  Event,
  catSvars = NULL,
  conSvars = NULL,
  y1 = T,
  y2 = T,
  y5 = T,
  medianCI = F
)
}
\arguments{
\item{Data}{A data.frame in which to interpret the variables}

\item{Stime}{This is the follow up time}

\item{Event}{The status indicator, normally 0=alive, 1=dead}

\item{catSvars}{A vector of cat variables}

\item{conSvars}{A vector of con variables}

\item{y1}{1-year survival rate (yes/no)}

\item{y2}{2-year survival rate(yes/no)}

\item{y5}{5-year survival rate(yes/no)}
}
\value{
A tibble of survival output
}
\description{
surv_multiuni output the table with general survival analysis result with Number of total patients,
Number of Events, Estimated Median, 1,2,5 year rate, HR (95\% Confidence Interval),P value This function only change the format of the output table.
}
\examples{
conSvars <- c("Age","KPS")
catSvars <- c("Gender")
surv_multiuni(D, "pfs", "rcensor", catSvars, conSvars,y2 = F, y5 = F, y1 = F)# \%>\% View
}
