% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_uni_cat.R
\name{surv_uni_cat}
\alias{surv_uni_cat}
\title{A complete table for univariable survival analysis}
\usage{
surv_uni_cat(
  Data,
  Stime,
  Event,
  Svar,
  month = 0,
  medianCI = F,
  y1 = T,
  y2 = T,
  y5 = T,
  AIC = F
)
}
\arguments{
\item{Stime}{This is the follow up time}

\item{Event}{The status indicator, normally 0=alive, 1=dead}

\item{Svar}{A vector of group}

\item{month}{month-rate of survival}

\item{y1}{1-year survival rate (yes/no)}

\item{y2}{2-year survival rate(yes/no)}

\item{y5}{5-year survival rate(yes/no)}

\item{AIC}{show AIC and C index (yes/no)}

\item{D}{A data.frame in which to interpret the variables}
}
\value{
A tibble of survival output
}
\description{
surv_uni_cat output the table with general survival analysis result with Number of total patients,
Number of Events, Estimated Median (95%CI), 1,2,5 year rate, HR (95\% Confidence Interval),P value, AIC and C index. This function only change the format of the output table.
}
\examples{
surv_uni_cat(Dat, "surv", "scensor", "DS-GPA")
}
