% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_uni_con.R
\name{surv_uni_con}
\alias{surv_uni_con}
\title{A complete table for univariable survival analysis}
\usage{
surv_uni_con(Data, Stime, Event, Svar)
}
\arguments{
\item{Stime}{This is the follow up time}

\item{Event}{The status indicator, normally 0=alive, 1=dead}

\item{Svar}{A vector of group}

\item{D}{A data.frame in which to interpret the variables}
}
\value{
A tibble of survival output
}
\description{
JS.uni output the table with general survival analysis result with Number of total patients,
Number of Events, Estimated Median, 1,2,5 year rate, HR (95\% Confidence Interval),P value, AIC and C index. This function only change the format of the output table.
}
\examples{
surv_uni_con(Dat, "surv", "scensor", "age")
}
